Account repo updated

+++ MD.Accountella.DL/Repositories/EntityCategoryRepository/EntityCategoryRepository.cs
@@ -12,16 +12,13 @@ namespace MD.Accountella.DL
     using System.Collections.Generic;
     using System.Linq;
     using System.Text;
-    internal class EntityCategoryRepository: IEntityCategoryRepository
+    internal class EntityCategoryRepository : IEntityCategoryRepository
     {
-        private readonly AccountellaDbContext _dbContext;
         private readonly IMongoCollection<EntityCategory> _catgColl;
         public EntityCategoryRepository(AccountellaDbContext dbContext)
         {
-            this._dbContext = dbContext;
-            var vv = new MongoCollectionSettings();
-            this._catgColl 
-                = this._dbContext.DB.GetCollection<EntityCategory>(nameof(EntityCategory));
+            this._catgColl
+                = dbContext.Db.GetCollection<EntityCategory>(nameof(EntityCategory));
         }
         public List<EntityCategory> GetCategories(string id)
         {
@@ -34,12 +31,12 @@ namespace MD.Accountella.DL
         public EntityCategory AddCategory(EntityCategory catgToAdd)
         {
             if (string.IsNullOrWhiteSpace(catgToAdd.Name))
-                throw new Exception("Category name must be provided");
+                throw new Exception("Name must be provided");
+
             var catgMatches = _catgColl.Find(catg =>
                         catg.Name.Equals(catgToAdd.Name)
                         && catg._parentId.Equals(catgToAdd._parentId)
                     );
-
             if (catgMatches.Any())
                 throw new Exception("Duplicate category is not allowed");
 
@@ -73,13 +70,22 @@ namespace MD.Accountella.DL
                 if (catgFound.First().IsReadOnly)
                     throw new InvalidOperationException("Unable to delete/remove readonly item");
 
+                //Checkig wheather same name is already exists with other or not
+                var catgMatches = _catgColl.Find(catg =>
+                            catg.Name.Equals(catgToUpdate.Name)
+                            && catg._parentId.Equals(catgToUpdate._parentId)
+                            && catg.Id != id
+                        );
+                if (catgMatches.Any())
+                    throw new Exception("Requested category name already exists");
+
                 catgToUpdate.Id = id;
                 catgToUpdate.IsReadOnly = catgFound.First().IsReadOnly;
                 catgToUpdate._parentId = catgFound.First()._parentId;
                 var result = _catgColl.ReplaceOne<EntityCategory>(catg => catg.Id.Equals(catgToUpdate.Id) && catg.IsReadOnly == false, catgToUpdate);
 
                 if (!result.IsModifiedCountAvailable)
-                    throw new SystemException("Unable to modify");                
+                    throw new SystemException("Unable to modify");
             }
             catch (Exception ex)
             {
@@ -88,18 +94,28 @@ namespace MD.Accountella.DL
         }
         public void RemoveCategory(string id)
         {
-            var catgFound = GetCategories(id);
+            try
+            {
+                var catgFound = GetCategories(id);
 
-            if (!catgFound.Any())
-                throw new KeyNotFoundException("Category requested for delete not found");
+                if (!catgFound.Any())
+                    throw new KeyNotFoundException("Category requested for delete not found");
 
-            var result = _catgColl.DeleteOne<EntityCategory>(catg => catg.Id.Equals(id) && catg.IsReadOnly == false);
+                if (catgFound.First().IsReadOnly)
+                    throw new Exception("Requested item is read only, hence cannot be deleted");
 
-            if (!result.IsAcknowledged)
-                throw new KeyNotFoundException("No matching category found with the provided key");
+                var result = _catgColl.DeleteOne<EntityCategory>(catg => catg.Id.Equals(id) && catg.IsReadOnly == false);
 
-            if (!(result.DeletedCount > 0))
-                throw new Exception("Unknown Error: Category not deleted");
+                if (!result.IsAcknowledged)
+                    throw new KeyNotFoundException("No matching category found with the provided key");
+
+                if (!(result.DeletedCount > 0))
+                    throw new Exception("Unknown Error: Category not deleted");
+            }
+            catch (Exception ex)
+            {
+                throw ex;
+            }
         }
     }
 }
